+++ components/usb/iot_usbh_modem/CHANGELOG.md
@@ -1,5 +1,10 @@
 # ChangeLog
 
+## v0.1.1 - 2023-02-13
+
+* Support IDF5.0
+* Change dependency iot_usbh_cdc to private
+
 ## v0.1.0 - 2023-02-02
 
 * initial version